# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Space Cargo Api",
        "version": "1.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Get a JWT access token via given credentials.",
                "operationId": "dd51a484b7fa0f97c7680fdbde14cf5f",
                "requestBody": {
                    "description": "Create customers object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "glowe"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "validation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/parcels/create": {
            "post": {
                "tags": [
                    "Parcels"
                ],
                "summary": "Create parcel",
                "operationId": "832cfda815dfe14aaa0cdc7119183a7b",
                "requestBody": {
                    "description": "Data for creating a new parcel",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "AC44BB13#TT"
                                    },
                                    "price": {
                                        "type": "numeric",
                                        "example": "12.33"
                                    },
                                    "quantity": {
                                        "type": "numeric",
                                        "example": "12"
                                    },
                                    "store_address": {
                                        "type": "string",
                                        "example": "test address"
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "test comment"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/parcels": {
            "get": {
                "tags": [
                    "Parcels"
                ],
                "summary": "Get a list of parcels",
                "operationId": "bf49128fb3aadfb9392953384eb9b6f5",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit the number of results",
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of parcels",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/parcels/{id}": {
            "get": {
                "tags": [
                    "Parcels"
                ],
                "summary": "Get a parcel",
                "operationId": "0d8d9cad98d1f70bc3bda1bf2471c46d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of parcels",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/parcels/update": {
            "put": {
                "tags": [
                    "Parcels"
                ],
                "summary": "Create parcel",
                "operationId": "fa51acd21a141231e6f2c7c6164ab40c",
                "requestBody": {
                    "description": "Data for creating a new parcel",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "numeric",
                                        "example": "1"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "AC44BB13#TT"
                                    },
                                    "price": {
                                        "type": "numeric",
                                        "example": "12.33"
                                    },
                                    "quantity": {
                                        "type": "numeric",
                                        "example": "12"
                                    },
                                    "store_address": {
                                        "type": "string",
                                        "example": "test address"
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "test comment"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}